+++ backend/internal/controller/repository/model/hello_world_model.go
@@ -1,0 +1,6 @@
+package model
+
+type HelloWorld struct {
+	Lang    string
+	Message string
+}
