+++ backend/internal/controller/repository/model/post_model.go
@@ -1,0 +1,13 @@
+package model
+
+import (
+	"gorm.io/gorm"
+)
+
+type Post struct {
+	UserID int    `json:"user_id"`
+	Title  string `json:"title"`
+	Body   string `json:"body"`
+
+	gorm.Model
+}
