+++ backend/internal/controller/repository/model/user_model.go
@@ -1,0 +1,11 @@
+package model
+
+import (
+	"gorm.io/gorm"
+)
+
+type User struct {
+	Name     string `json:"name"`
+	Password string `json:"password"`
+	gorm.Model
+}
