+++ backend/internal/controller/repository/user_repository.go
@@ -3,6 +3,7 @@ package repository
 import (
 	"errors"
 	"fmt"
+	"myapp/internal/controller/repository/model"
 	"myapp/internal/entity"
 	repositoryIF "myapp/internal/usecase/repository"
 
@@ -13,14 +14,6 @@ type UserRepository struct {
 	Conn *gorm.DB
 }
 
-// This struct is same as entity model
-// However define again for training
-type User struct {
-	Name     string `json:"name"`
-	Password string `json:"password"`
-	gorm.Model
-}
-
 func NewUserRepository(conn *gorm.DB) repositoryIF.UserRepository {
 	return &UserRepository{
 		Conn: conn,
@@ -28,7 +21,7 @@ func NewUserRepository(conn *gorm.DB) repositoryIF.UserRepository {
 }
 
 func (r *UserRepository) GetAll() ([]entity.User, error) {
-	var users []User
+	var users []model.User
 	result := r.Conn.Find(&users)
 	fmt.Printf("%+v\n", result)
 	fmt.Printf("%+v\n", users)
@@ -41,7 +34,7 @@ func (r *UserRepository) GetAll() ([]entity.User, error) {
 	return convertUserRepositoryModelToEntity(users), nil
 }
 
-func convertUserRepositoryModelToEntity(ps []User) []entity.User {
+func convertUserRepositoryModelToEntity(ps []model.User) []entity.User {
 	var users []entity.User
 
 	for _, p := range ps {
